refactor get records to parent class, each stream parses records differently
